Changes:
1.Added InformationManager:
Characters can use this manager to get user information (Full name, Id) and global information such as what characters are downloaded on the user device.

2. Added onUserEventOccurred method to IAliveAgent:
This method will be called once a user event will occur (such as when the user up/down vote a character, download a character etc..).